Refactored and working, might need balance
