+++ src/newcharacter.cpp
@@ -48,6 +48,7 @@
 #include "overmap_ui.h"
 #include "path_info.h"
 #include "pimpl.h"
+#include "player_difficulty.h"
 #include "profession.h"
 #include "proficiency.h"
 #include "recipe.h"
@@ -112,17 +113,7 @@ static bool isWide = false;
 #define COL_HEADER          c_white   // Captions, like "Profession items"
 #define COL_NOTE_MINOR      c_light_gray  // Just regular note
 
-// The point after which stats cost double
-static constexpr int HIGH_STAT = 12;
-
 static int skill_increment_cost( const Character &u, const skill_id &skill );
-
-enum class pool_type {
-    FREEFORM = 0,
-    ONE_POOL,
-    MULTI_POOL,
-    TRANSFER,
-};
 
 class tab_manager
 {
@@ -351,445 +342,6 @@ static std::string pools_to_string( const avatar &u, pool_type pool )
             return _( "Story Teller" );
     }
     return "If you see this, this is a bug";
-}
-
-// creates an npc with similar stats to an avatar for testing
-static void npc_from_avatar( const avatar &u, npc &dummy )
-{
-    // set stats
-    dummy.str_max = u.str_max;
-    dummy.dex_max = u.dex_max;
-    dummy.int_max = u.int_max;
-    dummy.per_max = u.per_max;
-    dummy.reset_stats();
-
-
-    // set skills
-    for( auto t : u.get_all_skills() ) {
-        dummy.set_skill_level( t.first, t.second.level() );
-    }
-
-    // set profession and hobbies
-    dummy.prof = u.prof;
-    dummy.hobbies = u.hobbies;
-
-    // set mutations
-    for( const trait_id &t : u.get_mutations( true ) ) {
-        dummy.set_mutation( t );
-    }
-
-    dummy.initialize( character_type::FULL_RANDOM );
-}
-
-static void reset_npc( Character &dummy )
-{
-    dummy.set_body();
-    dummy.normalize(); // In particular this clears martial arts style
-
-    // delete all worn items.
-    dummy.worn.clear();
-    dummy.calc_encumbrance();
-    dummy.inv->clear();
-    dummy.remove_weapon();
-    dummy.clear_mutations();
-
-    // Clear stomach and then eat a nutritious meal to normalize stomach
-    // contents (needs to happen before clear_morale).
-    dummy.stomach.empty();
-    dummy.guts.empty();
-    dummy.clear_vitamins();
-
-    // This sets HP to max, clears addictions and morale,
-    // and sets hunger, thirst, fatigue and such to zero
-    dummy.environmental_revert_effect();
-    // However, the above does not set stored kcal
-    dummy.set_stored_kcal( dummy.get_healthy_kcal() );
-
-    dummy.empty_skills();
-    dummy.martial_arts_data->clear_styles();
-    dummy.clear_morale();
-    dummy.clear_bionics();
-    dummy.activity.set_to_null();
-    dummy.reset_chargen_attributes();
-    dummy.set_pain( 0 );
-    dummy.reset_bonuses();
-    dummy.set_speed_base( 100 );
-    dummy.set_speed_bonus( 0 );
-    dummy.set_sleep_deprivation( 0 );
-    for( const proficiency_id &prof : dummy.known_proficiencies() ) {
-        dummy.lose_proficiency( prof, true );
-    }
-
-    // Reset cardio_acc to baseline
-    dummy.reset_cardio_acc();
-    // Restore all stamina and go to walk mode
-    dummy.set_stamina( dummy.get_stamina_max() );
-    dummy.reset_activity_level();
-
-    // Make sure we don't carry around weird effects.
-    dummy.clear_effects();
-
-    // Make stats nominal.
-    dummy.str_max = 8;
-    dummy.dex_max = 8;
-    dummy.int_max = 8;
-    dummy.per_max = 8;
-    dummy.set_str_bonus( 0 );
-    dummy.set_dex_bonus( 0 );
-    dummy.set_int_bonus( 0 );
-    dummy.set_per_bonus( 0 );
-}
-
-static double calc_armor_value( const Character &u, bodypart_id bp )
-{
-    // a low damage value to be concerned about early
-    // only concerned with bash damage since it is most
-    // prevalent early game
-
-
-
-    float armor_val = 0.0f;
-
-    // check any other items the character has on them
-    if( u.prof ) {
-        for( const item &i : u.prof->items( true, std::vector<trait_id>() ) ) {
-            armor_val += i.resist( damage_type::BASH, false, bp );
-        }
-    }
-
-    return armor_val;
-}
-
-static std::tuple<int, std::string> get_defense_difficulty( const Character &u )
-{
-    // figure out how survivable the character is.
-
-    // the percent margin between result bands
-    const float percent_band = 0.5f;
-
-    // guess at the ammount of armor an average clothed person would have
-    float base_armor = 2.0f;
-
-    // how much each portion is valued compared to the deviation of others
-    const float standard_movement = 1.0f;
-    const float difficult_movement = 2.0f;
-    const float dodge = 1.0f;
-    const float head_protection = 0.2f;
-    const float torso_protection = 0.4f;
-    const float arms_protection = 0.2f;
-    const float legs_protection = 0.2f;
-
-
-    // get a generic npc to compare against
-    npc average = npc();
-    reset_npc( average );
-
-    // calculate move cost on simple ground
-    int player_run_cost = u.run_cost( 100 );
-    int average_run_cost = average.run_cost( 100 );
-    float per = standard_movement * ( player_run_cost - average_run_cost ) / average_run_cost;
-
-    // calculate move cost on simple ground
-    player_run_cost = u.run_cost( 400 );
-    average_run_cost = average.run_cost( 400 );
-    per += difficult_movement * ( player_run_cost - average_run_cost ) / average_run_cost;
-
-    // calculate head armor
-    per += head_protection * ( calc_armor_value( u, bodypart_id( "head" ) ) - base_armor ) / base_armor;
-    // calculate torso armor
-    per += torso_protection * ( calc_armor_value( u,
-                                bodypart_id( "torso" ) ) - base_armor ) / base_armor;
-    // calculate arm armor
-    per += arms_protection * ( calc_armor_value( u,
-                               bodypart_id( "arm_r" ) ) - base_armor ) / base_armor;
-    // calculate leg armor
-    per += legs_protection * ( calc_armor_value( u,
-                               bodypart_id( "leg_r" ) ) - base_armor ) / base_armor;
-    // calculate dodge
-    per += dodge * ( u.get_dodge() - average.get_dodge() ) / average.get_dodge();
-
-    if( per <= -2 * percent_band ) {
-        return std::make_tuple( 5, string_format( "%2f Made Of Paper", per ) );
-    } else if( per <= -1 * percent_band ) {
-        return std::make_tuple( 4, string_format( "%2f Fragile", per ) );
-    } else if( per <= 0.0f ) {
-        return std::make_tuple( 3, string_format( "%2f Average", per ) );
-    } else if( per <= percent_band ) {
-        return std::make_tuple( 2, string_format( "%2f Tough", per ) );
-    } else if( per <= 2 * percent_band ) {
-        return std::make_tuple( 1, string_format( "%2f Hard To Kill", per ) );
-    }
-    return std::make_tuple( 1, string_format( "%2f Immortal", per ) );
-}
-
-static double calc_dps_value( const Character &u )
-{
-    // check against the big three
-    // efficient early weapons you can easily get access to
-    item early_piercing = item( "knife_combat" );
-    item early_cutting = item( "machete" );
-    item early_bashing = item( "bat" );
-
-    double baseline = std::max( u.weapon_value( early_piercing ),
-                                u.weapon_value( early_cutting ) );
-    baseline = std::max( baseline, u.weapon_value( early_bashing ) );
-
-
-
-    // check any other items the character has on them
-    if( u.prof ) {
-        for( const item &i : u.prof->items( true, std::vector<trait_id>() ) ) {
-            baseline = std::max( baseline, u.weapon_value( i ) );
-        }
-    }
-
-    return baseline;
-}
-
-static std::tuple<int, std::string> get_combat_difficulty( const Character &u )
-{
-    // figure out how good the player is at combat.
-    // get the best dps of a basic civilian with 3 scavengable weapons
-    // compare to the dps of this character with 1 of those 3 or their best weapon they spawn with
-
-    // the percent margin between result bands
-    const float percent_band = 0.5f;
-
-
-    // get a generic npc to compare against
-    npc average = npc();
-    reset_npc( average );
-
-    double npc_dps = calc_dps_value( average );
-    double player_dps = calc_dps_value( u );
-
-    float per = ( player_dps - npc_dps ) / npc_dps;
-
-    if( per <= -2 * percent_band ) {
-        return std::make_tuple( 5, string_format( "%2f Sitting Duck", per ) );
-    } else if( per <= -1 * percent_band ) {
-        return std::make_tuple( 4, string_format( "%2f Innefective", per ) );
-    } else if( per <= 0.0f ) {
-        return std::make_tuple( 3, string_format( "%2f Average", per ) );
-    } else if( per <= percent_band ) {
-        return std::make_tuple( 2, string_format( "%2f Scrapper", per ) );
-    } else if( per <= 2 * percent_band ) {
-        return std::make_tuple( 1, string_format( "%2f Trained Fighter", per ) );
-    }
-    return std::make_tuple( 1, string_format( "%2f Trained Killer", per ) );
-}
-
-static std::tuple<int, std::string> get_genetics_difficulty( const Character &u )
-{
-    // figure out how genetically advantaged the character is
-
-    // how many attributes the average character has
-    const int average_stats = 38;
-    // how much stats above HIGH_STAT penalize the player
-    const int high_stat_penalty = 2;
-    // how much traits are valued at
-    const int trait_value = 2;
-    // the percent margin between result bands
-    const float percent_band = 0.1f;
-
-
-    int genetics_total = u.str_max + u.dex_max + u.per_max + u.int_max;
-    genetics_total += std::max( 0, u.str_max - HIGH_STAT ) * high_stat_penalty;
-    genetics_total += std::max( 0, u.dex_max - HIGH_STAT ) * high_stat_penalty;
-    genetics_total += std::max( 0, u.per_max - HIGH_STAT ) * high_stat_penalty;
-    genetics_total += std::max( 0, u.int_max - HIGH_STAT ) * high_stat_penalty;
-
-    // each trait effects genetics slightly as well
-    for( const trait_id &trait : u.get_mutations( true ) ) {
-        genetics_total += trait->points > 0 ? trait_value : -1 * trait_value;
-    }
-
-    float per = static_cast<float>( genetics_total - average_stats ) / static_cast<float>
-                ( average_stats );
-
-    if( per <= -2 * percent_band ) {
-        return std::make_tuple( 5, string_format( "%2f Weak", per ) );
-    } else if( per <= -1 * percent_band ) {
-        return std::make_tuple( 4, string_format( "%2f Unimpressive", per ) );
-    } else if( per <= 0.0f ) {
-        return std::make_tuple( 3, string_format( "%2f Average", per ) );
-    } else if( per <= percent_band ) {
-        return std::make_tuple( 2, string_format( "%2f Impressive", per ) );
-    } else if( per <= 2 * percent_band ) {
-        return std::make_tuple( 1, string_format( "%2f Powerful", per ) );
-    }
-    return std::make_tuple( 1, string_format( "%2f Superhuman", per ) );
-}
-
-static std::tuple<int, std::string> get_expertise_difficulty( const Character &u )
-{
-    // a bit extra over multipool since you get 2 points per in multi
-    const int average_skill_ranks = 4;
-    const float per_bands = 0.6f;
-
-    // how much each proficiency is valued compared to a skill point
-    const int proficiency_value = 2;
-
-    // how much each portion is valued compared to the deviation of others
-    const float skill_weighting = 1.0f;
-    const float reading_weighting = 2.0f;
-    const float learn_weighting = 2.0f;
-    const float focus_weighting = 2.0f;
-
-
-    // sum player skills and proficiencies
-    int player_skills = 0;
-    // every skill point is worth 1 point of value
-    for( const auto &t : u.get_all_skills() ) {
-        // combat skills will be handled in offence
-        if( !t.first->is_combat_skill() ) {
-            player_skills += t.second.level();
-        }
-    }
-    // every proficiency is worth about the value of 2 skill points
-    for( const proficiency_id &t : u._proficiencies->known_profs() ) {
-        player_skills += proficiency_value;
-    }
-
-
-    // get a generic npc to compare against
-    npc average = npc();
-    reset_npc( average );
-
-    // skills and professions
-    float per = skill_weighting * static_cast<float>( player_skills - average_skill_ranks ) /
-                static_cast<float>( average_skill_ranks );
-
-    // focus
-    per += focus_weighting * static_cast<float>( u.calc_focus_equilibrium(
-                true ) - average.calc_focus_equilibrium(
-                true ) ) / static_cast<float>( average.calc_focus_equilibrium( true ) );
-
-    // reading speed negative is good
-    per -= reading_weighting * static_cast<float>( u.read_speed() - average.read_speed() ) /
-           static_cast<float>( average.read_speed() );
-
-    // how much each point of experience is worth to your character
-    per += learn_weighting * static_cast<float>( u.adjust_for_focus( 100 ) -
-            average.adjust_for_focus( 100 ) ) / static_cast<float>( average.adjust_for_focus( 100 ) );
-
-
-
-    if( per <= -1 * per_bands ) {
-        return std::make_tuple( 5, string_format( "%2f Hopeless", per ) );
-    } else if( per <= 0.0f ) {
-        return std::make_tuple( 4, string_format( "%2f Mediocre", per ) );
-    } else if( per <= per_bands ) {
-        return std::make_tuple( 3, string_format( "%2f Average", per ) );
-    } else if( per <= 2 * per_bands ) {
-        return std::make_tuple( 2, string_format( "%2f Skilled", per ) );
-    }
-    return std::make_tuple( 1, string_format( "%2f Brilliant", per ) );
-}
-
-
-static int calc_social_value( const Character &u, npc &compare )
-{
-    // weighting skill and underlying values equally
-
-    int social = std::max( u.intimidation(), u.persuade_skill() );
-    int lying = u.lie_skill();
-
-    compare.form_opinion( u );
-
-    int oppinion = compare.op_of_u.trust;
-    oppinion -= compare.op_of_u.anger;
-    oppinion += compare.op_of_u.fear;
-    oppinion += compare.op_of_u.value;
-
-    return social + lying + oppinion;
-}
-
-static std::tuple<int, std::string> get_social_difficulty( const Character &u )
-{
-
-
-    npc average = npc();
-    reset_npc( average );
-
-    // compare the characters social value to an average npc
-    int player_val = calc_social_value( u, average );
-    int average_val = calc_social_value( average, average );
-
-
-    float per = static_cast<float>( player_val - average_val ) / static_cast<float>
-                ( average_val );
-    if( per <= -0.6f ) {
-        return std::make_tuple( 5, string_format( "%2f Sheltered", per ) );
-    } else if( per <= 0.0f ) {
-        return std::make_tuple( 4, string_format( "%2f Awkward", per ) );
-    } else if( per <= 0.6f ) {
-        return std::make_tuple( 3, string_format( "%2f Average", per ) );
-    } else if( per <= 1.2f ) {
-        return std::make_tuple( 2, string_format( "%2f Social", per ) );
-    }
-    return std::make_tuple( 1, string_format( "%2f Charasmatic", per ) );
-}
-
-static std::string difficulty_to_string( const avatar &u )
-{
-    float off_diff = 3.0f;
-    float def_diff = 3.0f;
-    float cft_diff = 3.0f;
-    float wld_diff = 3.0f;
-    float scl_diff = 3.0f;
-    int off_cnt = 0;
-    int def_cnt = 0;
-    int cft_cnt = 0;
-    int wld_cnt = 0;
-    int scl_cnt = 0;
-
-    int off = std::round( off_diff / ( off_cnt > 0 ? off_cnt : 1 ) );
-    int def = std::round( def_diff / ( def_cnt > 0 ? def_cnt : 1 ) );
-    int cft = std::round( cft_diff / ( cft_cnt > 0 ? cft_cnt : 1 ) );
-    int wld = std::round( wld_diff / ( wld_cnt > 0 ? wld_cnt : 1 ) );
-    int scl = std::round( scl_diff / ( scl_cnt > 0 ? scl_cnt : 1 ) );
-
-    auto diff_colr = []( int diff ) {
-        switch( diff ) {
-            case 1:
-                return "light_green";
-            case 2:
-                return "light_cyan";
-            case 4:
-                return "brown";
-            case 5:
-                return "light_red";
-            default:
-                return "yellow";
-        }
-    };
-
-    auto diff_desc = []( int diff ) {
-        return std::string( "test" );
-    };
-
-    // apply professions, hobbies, etc for testing
-
-    // make a faux avatar that can have the effects of creation applied to it
-    npc n = npc();
-    reset_npc( n );
-    npc_from_avatar( u, n );
-
-    std::tuple<int, std::string> genetics = get_genetics_difficulty( n );
-    std::tuple<int, std::string> socials = get_social_difficulty( n );
-    std::tuple<int, std::string> expertise = get_expertise_difficulty( n );
-    std::tuple<int, std::string> combat = get_combat_difficulty( n );
-    std::tuple<int, std::string> defense = get_defense_difficulty( n );
-
-
-    return string_format( "%s |  %s: <color_%s>%s</color>  %s: <color_%s>%s</color>  %s: <color_%s>%s</color>  %s: <color_%s>%s</color>  %s: <color_%s>%s</color>",
-                          _( "Summary" ),
-                          _( "Genetics" ), diff_colr( std::get<0>( genetics ) ), std::get<1>( genetics ),
-                          _( "Offense" ), diff_colr( std::get<0>( combat ) ), std::get<1>( combat ),
-                          _( "Defense" ), diff_colr( std::get<0>( defense ) ), std::get<1>( defense ),
-                          _( "Expertise" ), diff_colr( std::get<0>( expertise ) ), std::get<1>( expertise ),
-                          _( "Social" ), diff_colr( std::get<0>( socials ) ), std::get<1>( socials ) );
 }
 
 static void set_points( tab_manager &tabs, avatar &u, pool_type & );
@@ -1379,7 +931,8 @@ static void draw_points( const catacurses::window &w, pool_type pool, const avat
             mvwprintz( w, point( pMsg_length + 2, 3 ), c_green, " (+%d)", std::abs( netPointCost ) );
         }
     }
-    print_colored_text( w, point( 2, 4 ), color, c_light_gray, difficulty_to_string( u ) );
+    print_colored_text( w, point( 2, 4 ), color, c_light_gray,
+                        player_difficulty::getInstance().difficulty_to_string( u ) );
 }
 
 template <class Compare>
